:recycle: refactor ApplicationCommands to Commands
