+++ package.json
@@ -49,6 +49,7 @@
     "mocha": "9.0.1",
     "nodemon": "2.0.12",
     "prettier": "2.4.1",
+    "prettier-plugin-organize-imports": "^2.3.4",
     "ts-mockito": "2.6.1",
     "ts-node": "10.0.0",
     "tsc-alias": "1.3.9",
