+++ src/controllers/interaction-controller.ts
@@ -1,20 +1,22 @@
-import { injectable } from 'inversify';
+import { Command } from '@commands/command';
+import commandDefinitions from '@commands/definitions';
 import { Controller } from '@controllers/controller';
-import { REST } from '@discordjs/rest';
-import { Routes } from 'discord-api-types/v9';
-import interactions from '@commands/interactions';
-import { CommandInteraction, Interaction } from 'discord.js';
-import { ApplicationCommand } from '@commands/interactions/application-command';
+import { REST, RouteLike } from '@discordjs/rest';
+import { CommandError } from '@models/commands/command-error';
+import { CommandValidationError } from '@models/commands/command-validation-error';
 import container from '@src/inversify.config';
-import { ApplicationCommandError } from '@models/application-command-error';
+import { TYPES } from '@src/types';
+import { Routes } from 'discord-api-types/v9';
+import { CommandInteraction, Interaction } from 'discord.js';
+import { injectable } from 'inversify';
 
 @injectable()
 export class InteractionController extends Controller {
     public async registerApplicationCommands(): Promise<number> {
         // Set up requests
         const jsonPayload = [];
-        for (let i = 0; i < interactions.length; i++) {
-            jsonPayload.push(interactions[i].toJSON());
+        for (let i = 0; i < commandDefinitions.length; i++) {
+            jsonPayload.push(commandDefinitions[i].toJSON());
         }
         const rest = new REST({ version: '9' }).setToken(this.token);
 
@@ -22,12 +24,29 @@ export class InteractionController extends Controller {
         try {
             this.logger.debug('Refreshing application commands...');
 
-            await rest.put(Routes.applicationCommands(this.clientId), {
-                body: interactions,
+            // Delete any global commands
+            rest.get(Routes.applicationCommands(this.clientId)).then((data: any) => {
+                const promises = [];
+                for (const command of data) {
+                    const deleteUrl = `${Routes.applicationCommands(this.clientId)}/${command.id}`;
+                    promises.push(rest.delete(<RouteLike>deleteUrl));
+                }
+                return Promise.all(promises);
             });
 
+            // Register guild commands
+            await rest.put(
+                Routes.applicationGuildCommands(
+                    this.clientId,
+                    container.get<string>(TYPES.GuildId)
+                ),
+                {
+                    body: commandDefinitions,
+                }
+            );
+
             this.logger.debug('Successfully refreshed application commands.');
-            return Promise.resolve(interactions.length);
+            return Promise.resolve(commandDefinitions.length);
         } catch (error) {
             this.logger.error(
                 'Failed to register application commands: ',
@@ -48,30 +67,70 @@ export class InteractionController extends Controller {
     }
 
     private async handleApplicationCommand(interaction: CommandInteraction): Promise<void> {
-        try {
-            const applicationCommandName =
-                'Application' +
-                interaction.commandName.charAt(0).toUpperCase() +
-                interaction.commandName.slice(1);
-            const applicationCommand = container.get(applicationCommandName) as ApplicationCommand;
-            await applicationCommand
-                .run(interaction)
-                .then((result) => {
-                    this.logger.info(
-                        `Interaction ID ${interaction.id}: Application Command ${
-                            interaction.commandName
-                        } was ${result.executed ? '' : 'not'} executed: ${result.message}`
-                    );
-                })
-                .catch((error: ApplicationCommandError) => {
-                    this.logger.error(error.internalMessage, this.logger.prettyError(error.error));
-                    interaction.reply({
-                        content: error.userMessage,
-                        ephemeral: true,
-                    });
+        // Match the Command by name
+        const applicationCommandName =
+            interaction.commandName.charAt(0).toUpperCase() + interaction.commandName.slice(1);
+        const applicationCommand = container.get(applicationCommandName) as Command;
+
+        // Validate the inputs
+        let valid = true;
+        await applicationCommand
+            .validateOptions(interaction.options)
+            .catch((error: CommandValidationError) => {
+                // User input is not valid
+                this.logger.info(error.internalMessage);
+                interaction.reply({
+                    content: error.userMessage,
+                    ephemeral: true,
                 });
-        } catch (error) {
-            this.logger.error(`Failed: `, this.logger.prettyError(error));
-        }
+                valid = false;
+                // TODO: It continues to run the command despite going in here????
+            })
+            .catch((error) => {
+                // Unexpected error
+                this.logger.error(
+                    `Interaction ID ${interaction.id}: Application Command ${interaction.commandName} failed unexpectedly while validating options: `,
+                    this.logger.prettyError(error)
+                );
+                valid = false;
+            });
+        if (!valid) return Promise.resolve();
+
+        // Run the command
+        await applicationCommand
+            .run(interaction)
+            .then((result) => {
+                this.logger.info(
+                    `Interaction ID ${interaction.id}: Application Command ${
+                        interaction.commandName
+                    } was ${result.executed ? '' : 'not'} executed: ${result.message}`
+                );
+            })
+            .catch((error: CommandValidationError) => {
+                // Further user input validation failed
+                this.logger.info(error.internalMessage);
+                interaction.reply({
+                    content: error.userMessage,
+                    ephemeral: true,
+                });
+            })
+            .catch((error: CommandError) => {
+                // Command failed
+                this.logger.error(
+                    error.internalMessage,
+                    error.error ? this.logger.prettyError(error.error) : null
+                );
+                interaction.reply({
+                    content: error.userMessage,
+                    ephemeral: true,
+                });
+            })
+            .catch((error) => {
+                // Command failed unexpectedly
+                this.logger.error(
+                    `Interaction ID ${interaction.id}: Application Command ${interaction.commandName} failed unexpectedly while executing: `,
+                    this.logger.prettyError(error)
+                );
+            });
     }
 }
