+++ src/models/commands/command-validation-error.ts
@@ -1,0 +1,9 @@
+export class CommandValidationError {
+    internalMessage: string;
+    userMessage: string;
+
+    constructor(internalMessage: string, userMessage: string) {
+        this.internalMessage = internalMessage;
+        this.userMessage = userMessage;
+    }
+}
