+++ src/models/data/character-schema.ts
