+++ src/models/data/character.ts
