+++ src/providers/string-provider.ts
@@ -1,18 +1,15 @@
-import { injectable } from 'inversify';
 import { Provider } from '@src/providers/provider';
+import { injectable } from 'inversify';
 import { String } from 'typescript-string-operations';
 
 @injectable()
 export class StringProvider extends Provider {
     public async get(key: string, parameters?: string[]): Promise<string> {
-        // Get the string value from the Redis store
         let result = await this.redisClient.get('STRINGS.' + key);
         if (!result) {
             // Since strings are used in message sending, make sure that they are not null
             this.logger.warn(`Failed to get string with key STRINGS.${key}`);
-            return Promise.resolve(
-                `\`Internal Error: Cannot resolve key ${key}.\`\nYour command was still successful.`
-            );
+            return `\`Internal Error: Cannot resolve key ${key}.\`\n`;
         }
 
         // Append the parameters, if there are any
@@ -25,12 +22,9 @@ export class StringProvider extends Provider {
                         parameters.length
                     } parameters: ${JSON.stringify(parameters)}`
                 );
-                return Promise.resolve(
-                    `\`Internal Error: Cannot resolve parameters for key ${key}\`\nYour command was still successful.`
-                );
+                return `\`Internal Error: Cannot resolve parameters for key ${key}\``;
             }
         }
-
-        return Promise.resolve(result);
+        return result;
     }
 }
