+++ src/services/index.ts
@@ -1,6 +1,5 @@
-export * from '@services/permission-service';
-export * from '@services/message-service';
-export * from '@services/command-service';
 export * from '@services/channel-service';
-export * from '@services/user-service';
 export * from '@services/helper-service';
+export * from '@services/message-service';
+export * from '@services/permission-service';
+export * from '@services/user-service';
